// src/findjobnu-api/models/FindjobnuServiceDTOsProfileDto.ts
export interface FindjobnuServiceDTOsProfileDto {
    /**
     * 
     * @type {number}
     * @memberof FindjobnuServiceDTOsProfileDto
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof FindjobnuServiceDTOsProfileDto
     */
    userId?: string | null;
    /**
     * 
     * @type {Date}
     * @memberof FindjobnuServiceDTOsProfileDto
     */
    lastUpdatedAt?: Date | null;
    /**
     * 
     * @type {Date}
     * @memberof FindjobnuServiceDTOsProfileDto
     */
    createdAt?: Date;
    /**
     * 
     * @type {Array<string>}
     * @memberof FindjobnuServiceDTOsProfileDto
     */
    savedJobPosts?: Array<string> | null;
    /**
     * 
     * @type {Array<string>}
     * @memberof FindjobnuServiceDTOsProfileDto
     */
    keywords?: Array<string> | null;
    /**
     * 
     * @type {FindjobnuServiceDTOsBasicInfoDto}
     * @memberof FindjobnuServiceDTOsProfileDto
     */
    basicInfo?: FindjobnuServiceDTOsBasicInfoDto;
    /**
     * 
     * @type {Array<FindjobnuServiceDTOsExperienceDto>}
     * @memberof FindjobnuServiceDTOsProfileDto
     */
    experiences?: Array<FindjobnuServiceDTOsExperienceDto> | null;
    /**
     * 
     * @type {Array<FindjobnuServiceDTOsEducationDto>}
     * @memberof FindjobnuServiceDTOsProfileDto
     */
    educations?: Array<FindjobnuServiceDTOsEducationDto> | null;
    /**
     * 
     * @type {Array<FindjobnuServiceDTOsInterestDto>}
     * @memberof FindjobnuServiceDTOsProfileDto
     */
    interests?: Array<FindjobnuServiceDTOsInterestDto> | null;
    /**
     * 
     * @type {Array<FindjobnuServiceDTOsAccomplishmentDto>}
     * @memberof FindjobnuServiceDTOsProfileDto
     */
    accomplishments?: Array<FindjobnuServiceDTOsAccomplishmentDto> | null;
    /**
     * 
     * @type {Array<FindjobnuServiceDTOsContactDto>}
     * @memberof FindjobnuServiceDTOsProfileDto
     */
    contacts?: Array<FindjobnuServiceDTOsContactDto> | null;
    /**
     * 
     * @type {Array<FindjobnuServiceDTOsSkillDto>}
     * @memberof FindjobnuServiceDTOsProfileDto
     */
    skills?: Array<FindjobnuServiceDTOsSkillDto> | null;
    /**
     * 
     * @type {boolean}
     * @memberof FindjobnuServiceDTOsProfileDto
     */
    hasJobAgent?: boolean;
    /**
     * 
     * @type {FindjobnuServiceDTOsJobAgentDto}
     * @memberof FindjobnuServiceDTOsProfileDto
     */
    jobAgent?: FindjobnuServiceDTOsJobAgentDto;
}